backend/requirements.txt
Pillow==9.2.0
django-colorfield==0.7.1
django-jazzmin==2.5.0
django==3.2.14
psycopg2-binary==2.9.3
python-dotenv==0.20.0
djangorestframework==3.13.1
django-filter==22.1
djoser==2.1.0
drf-extra-fields==3.4.0
